+++ Toolkit/Public/Get-RscMssqlInstance.ps1
@@ -14,11 +14,14 @@ function Get-RscMssqlInstance{
     .PARAMETER List
     Uses the latest recovery point date and time that Rubrik has for a database
 
-    .PARAMETER RscHost
-    RscHost object retrieved via Get-RscHost
+    .PARAMETER HostName
+    Host name of the SQL Server Instance
+
+    .PARAMETER WindowsClusterName
+    Windows Cluster Name of the SQL Server Instance
 
     .PARAMETER InstanceName
-    SQL Server Instance Name
+    SQL Server Instance Name. If not provided, we default to MSSQLSERVER
     
     .PARAMETER Id
     Used to return a specific SQL Server Instance based on the Id assigned inside of Rubrik
@@ -36,16 +39,19 @@ function Get-RscMssqlInstance{
     .EXAMPLE
     Returns information about the default instance of SQL on a specific host
     $HostName = "rp-sql19s-001.demo.rubrik.com"
-    $RscHost = Get-RscHost -Name $HostName -OsType Windows
-    $RscMssqlInstance = Get-RscMssqlInstance -RscHost $RscHost
+    $RscMssqlInstance = Get-RscMssqlInstance -HostName $HostName
     
     .EXAMPLE
     Returns information about a specific instance of SQL on a specific host
     $HostName = "rp-sql19s-001.demo.rubrik.com"
-    $RscHost = Get-RscHost -Name $HostName -OsType Windows
-    $RscMssqlInstance = Get-RscMssqlInstance -RscHost $RscHost -InstanceName DEV01
+    $RscMssqlInstance = Get-RscMssqlInstance -HostName $HostName -InstanceName DEV01
+
+    .EXAMPLE
+    Returns information about a specific instance of SQL on a Windows Cluster
+    $WindowsClusterName = "rp-winfcsql"
+    $RscMssqlInstance = Get-RscMssqlInstance -WindowsClusterName $WindowsClusterName -InstanceName DEV01
     
-        .EXAMPLE
+    .EXAMPLE
     Return a RscMssqlInstance Object based on a specific MssqlInstance Id
     Get-RscMssqlInstance -Id "86da734b-2fee-4fdc-bdc8-a73ab5648f" 
     #>
@@ -54,36 +60,28 @@ function Get-RscMssqlInstance{
         DefaultParameterSetName = "List"
     )]
     Param(
-        [Parameter(
-            ParameterSetName = "List",
-            Mandatory = $false
-        )][Switch]$List,
-
-        [Parameter(
-            ParameterSetName = "RscHost",
-            Mandatory = $false
-        )][RubrikSecurityCloud.Types.PhysicalHost]$RscHost,
-
-        [Parameter(
-            ParameterSetName = "RscHost",
-            Mandatory = $false
-        )][String]$InstanceName = "MSSQLSERVER",
+        [Parameter(ParameterSetName = "List",Mandatory = $false)]
+        [Switch]$List,
         
-        [Parameter(
-            ParameterSetName = "Id",
-            Mandatory = $false, 
-            ValueFromPipeline = $false
-        )][String]$Id,
-       
-        [Parameter(
-            Mandatory = $false
-        )][RubrikSecurityCloud.Types.Cluster]$RscCluster,
+        [Parameter(ParameterSetName = "ByHostName", Mandatory = $true)]
+        [String]$HostName,
+        
+        [Parameter(ParameterSetName = "ByWindowsClusterName", Mandatory = $true)]
+        [String]$WindowsClusterName,
 
-        #  Common parameter to all parameter sets:
-        [Parameter(
-            Mandatory = $false, 
-            ValueFromPipeline = $false
-        )][Switch]$Detail
+        [Parameter(ParameterSetName = "ByHostName", Mandatory = $false)]
+        [Parameter(ParameterSetName = "ByWindowsClusterName", Mandatory = $false)]
+        [String]$InstanceName = "MSSQLSERVER",
+        
+        [Parameter(ParameterSetName = "Id", Mandatory = $true)]
+        [String]$Id,
+       
+        [Parameter(Mandatory = $false)]
+        [RubrikSecurityCloud.Types.Cluster]$RscCluster,
+
+        # Common parameter to all parameter sets:
+        [Parameter(Mandatory = $false, ValueFromPipeline = $false)]
+        [Switch]$Detail
     )
     Process {
         # Determine field profile:
@@ -91,55 +89,75 @@ function Get-RscMssqlInstance{
         if ( $Detail -eq $true ) {
             $fieldProfile = "DETAIL"
         }
-        Write-Debug "-Running Get-RscMssqlInstance"
-
+        Write-Verbose "-Running Get-RscMssqlInstance"
         #region Create Query
-        switch ( $PSCmdlet.ParameterSetName){
+        switch($PSCmdlet.ParameterSetName){
             "List"{
-                Write-Debug "-  Creating List Query"
+                Write-Verbose "-  Creating List Query"
                 $query = New-RscQueryMssql -Op TopLevelDescendants -FieldProfile $fieldProfile
-                $query.Var.filter = @()
-                $query.Var.typeFilter = "PhysicalHost"
+                $query.Var.typeFilter = @()
+                $query.Var.typeFilter += "PhysicalHost"
+                $query.Var.typeFilter += "WindowsCluster"
             }
-            "RscHost"{
-                Write-Debug "-  Creating Host Query"
-                $query = New-RscQueryMssql -Op TopLevelDescendants -FieldProfile $fieldProfile
+            {($_ -eq "ByHostName") -or ($_ -eq "ByWindowsClusterName")}{
+                $Name = ""
+                if($HostName){$Name = $HostName}
+                if($WindowsClusterName){$Name = $WindowsClusterName}
+                $query = New-RscQueryMssql -Operation TopLevelDescendants -FieldProfile $fieldProfile 
                 $query.Var.filter = @()
+                $query.Field.Nodes[3].PhysicalChildConnection.Nodes[5].Cluster = New-Object RubrikSecurityCloud.Types.Cluster
+                $query.Field.Nodes[3].PhysicalChildConnection.Nodes[5].Cluster.SelectForRetrieval()
+                $query.Field.Nodes[3].PhysicalChildConnection.Nodes[5].PhysicalPath = New-Object RubrikSecurityCloud.Types.PathNode
+                $query.Field.Nodes[3].PhysicalChildConnection.Nodes[5].PhysicalPath.SelectForRetrieval()
+    
+                $query.field.Nodes[4].LogicalChildConnection = New-Object RubrikSecurityCloud.Types.WindowsClusterLogicalChildTypeConnection
+                $query.field.Nodes[4].LogicalChildConnection.SelectForRetrieval()
+                $query.Field.Nodes[4].LogicalChildConnection.Nodes[0].PhysicalPath = New-Object RubrikSecurityCloud.Types.PathNode
+                $query.Field.Nodes[4].LogicalChildConnection.Nodes[0].PhysicalPath.SelectForRetrieval()
+    
                 $nameFilter = New-Object -TypeName RubrikSecurityCloud.Types.Filter
                 $nameFilter.Field = [RubrikSecurityCloud.Types.HierarchyFilterField]::NAME_EXACT_MATCH
-                $nameFilter.texts = $RscHost.Name
+                $nameFilter.texts = $Name
                 $query.Var.filter += $nameFilter
             }
-            "Id"  {
-                Write-Debug "-  Creating Id Query"
-                $query = New-RscQueryMssql -Op Instance -FieldProfile $fieldProfile -AddField PhysicalPath
+            "Id"{
+                Write-Verbose "-  Creating Id Query"
+                $query = New-RscQueryMssql -Operation Instance -FieldProfile $fieldProfile
+                $query.Field.PhysicalPath = New-Object RubrikSecurityCloud.Types.PathNode
+                $query.Field.PhysicalPath.SelectForRetrieval()
                 $query.Var.filter = @()
                 $query.Var.fid = $id
             }
         }
-
-        if($RscCluster){
-            Write-Debug "-  Creating Cluster Filter"
+        
+        if ($PSBoundParameters.ContainsKey('RscCluster')) {
+            Write-Verbose "-  Creating Cluster Filter"
             $clusterFilter = New-Object -TypeName RubrikSecurityCloud.Types.Filter
             $clusterFilter.Field = [RubrikSecurityCloud.Types.HierarchyFilterField]::CLUSTER_ID
             $clusterFilter.texts = $RscCluster.Id
             $query.Var.filter += $clusterFilter
         }
         #endregion
-
-        $result = $query.Invoke()
         
-        switch ( $PSCmdlet.ParameterSetName ){
-            "RscHost" {
-                Write-Debug "-  Filtering results based on Instance Name"
-                $result = $result.Nodes.PhysicalChildConnection.Nodes | Where-Object {$_.Name -eq $InstanceName}
-                $result
+        $results = $query.Invoke()
+
+        switch($PSCmdlet.ParameterSetName){
+            "List"{
+                $results.Nodes
+            }
+            {($_ -eq "ByHostName") -or ($_ -eq "ByWindowsClusterName")}{
+                switch($results.nodes.objectType){
+                    "PHYSICAL_HOST"{
+                        $results = $results.Nodes.PhysicalChildConnection.Nodes | Where-Object {$_.Name -eq $InstanceName}
+                    }
+                    "WINDOWS_CLUSTER"{
+                        $results = $results.Nodes.LogicalChildConnection.Nodes | Where-Object {$_.Name -eq $InstanceName}
+                    }
+                }
+                [RubrikSecurityCloud.Types.MssqlInstance]$results
             }
             "Id"{
-                $result
-            }
-            default{
-                $result.Nodes.PhysicalChildConnection.Nodes
+                $results
             }
         }
     }